Move models to data layer
